requirements.txt
Django==3.0
djangorestframework==3.11.0
mysqlclient==1.3.13
uvicorn